feat: migrate money to Long cents with tested business rules
- All monetary fields migrated Double -> Long (cents) in 5 entities:
  ProdutoEntity, ProdutoVariacaoEntity, PedidoEntity, ItemPedidoEntity,
  PagamentoEntity. Coordinates (lat/lng) kept as Double.
- AppDatabase version 1 -> 2 with fallbackToDestructiveMigration
- Domain models (Models.kt) migrated: precoBase, precoAdicional,
  precoUnitario, subtotal (computed), totalProdutos/Frete/Desconto/Final,
  Pagamento.valor
- getReceitaPeriodo: Double -> Long through DAO, repository interface
  and implementation
- CurrencyFormatter: manual implementation without NumberFormat/NBSP;
  Long.toMoedaBR(), Long.toMoedaBRInput(), String.toCentavos()
- CalcularPedido object: subtotalItem, totalProdutos, totalFinal,
  totalComJuros, valorParcela, valorPendente, freteCalculado
- ProdutoFormScreen/ViewModel, NovoPedidoViewModel: use Long cents
- JUnit4 dependency added, 11 new tests (17 total), all passing
- BUILD SUCCESSFUL

+++ app/src/main/java/com/confeitaria/gestao/data/local/database/entity/PagamentoEntity.kt
@@ -14,7 +14,7 @@ import androidx.room.PrimaryKey
 data class PagamentoEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val pedidoId: Long,
-    val valor: Double,
+    val valor: Long,
     val formaPagamento: String,
     val dataPagamento: Long = System.currentTimeMillis(),
     val confirmado: Boolean = false
